requirements.txt
pytest>=3.0
numpy>=1.20.1
matplotlib.pyplot
sklearn.model_selection
pandas